tests ~ refactor `createTestFn(PATH)` to probe the stack if no PATH is given
- PATH is now optional, probing the stack for the original caller info when not given
- export a "hand-free" `test()` function from '$shared.ts' which allows easy use of the expanded testing function

# why

Probing the stack gives information about the caller which can be used to give more
detailed information about the specific test, namely caller file path, as well as line and
character numbers of the specific call. There is some overhead, though tiny, per test call
for this information.

Additionally, refactoring `createTestFn(...)` to allow a missing/null PATH and
automatically probing the stack when PATH is missing creates a function requiring less
user configuration. Directly supplying the `test()` function makes it essentially a
"hands-free" drop-in for the standard Deno `test()` function.