docs ~ fix `pathIntoURL()` == uses 'file semantics' for resolution of path(s) from base
## notes

'URL semantics' for path resolution depend on whether the base URL has a final path
separator or not. If it does, then the path is resolved relative to the base URL's full
path. If not, then the path is resolved relative to the base URL's parent directory.

'File semantics' for path resolution always resolves the path relative to the base URL's
full path, regardless of whether the base URL has a final path separator or not.